+++ resources/views/layouts/navigation.blade.php
@@ -173,10 +173,13 @@
 </header>
 <header id="topNavigation">
 		<form>
-			<i class="icon" data-feather="search"></i>
-			<input autocomplete="off" data-toggle="popover" data-placement="bottom" onkeyup="ajaxSearch()" class="search-input" type="text" name="searchQ" placeholder="Ieškoti transporto priemonių, žmonių, maršrutų, klientų, pavedimų ir kitos informacijos">
+			<i class="icon icon-white" data-feather="search"></i>
+			<input autocomplete="off" data-toggle="popover" data-placement="bottom" onkeyup="ajaxSearch()" class="search-input" type="text" name="searchQ" placeholder="Paieška">
 		</form>
-		<span style="color: #fff;">{{ Auth::user()->name }}</span>
+		<span style="color: #fff;">
+      <span class="icon icon-white mr-1" data-feather="user"></span>
+      {{ Auth::user()->name }}
+    </span>
 </header>
 	<div class="sideNavigation">
     <div id="NavExpandTriggerBox">
@@ -186,22 +189,22 @@
     </div>
 		<div class="items">
 			<a href="{{ route('home') }}" class=" @if (Route::currentRouteName() == "home") active @endif ">
-				<span class="icon" data-feather="home"></span> <span class="ExpandableItem">Dokumentacija</span>
+				<span class="icon" data-feather="home" id="navIcon"></span> <span class="ExpandableItem">Dokumentacija</span>
 			</a>
       <a href="{{ route('drivers.index') }}" class=" @if (Route::currentRouteName() == "drivers.index") active @endif ">
-        <span class="icon" data-feather="users"></span> <span class="ExpandableItem">Vairuotojai</span>
+        <span class="icon" data-feather="users" id="navIcon"></span> <span class="ExpandableItem">Vairuotojai</span>
       </a>
       <a href="{{ route('routes.index') }}" class=" @if (Route::currentRouteName() == "routes.index") active @endif ">
-        <span class="icon" data-feather="map"></span> <span class="ExpandableItem">Maršrutai</span>
+        <span class="icon" data-feather="map" id="navIcon"></span> <span class="ExpandableItem">Maršrutai</span>
       </a>
 			<a href="{{ route('repairs.index') }}" class=" @if (Route::currentRouteName() == "repairs.index") active @endif ">
-				<span class="icon" data-feather="activity"></span> <span class="ExpandableItem">Gedimai</span>
+				<span class="icon" data-feather="activity" id="navIcon"></span> <span class="ExpandableItem">Gedimai</span>
 			</a>
 			<a href="{{ route('insurance.index') }}">
-				<span class="icon" data-feather="briefcase"></span> <span class="ExpandableItem">Draudimas</span>
+				<span class="icon" data-feather="briefcase" id="navIcon"></span> <span class="ExpandableItem">Draudimas</span>
 			</a>
 			<a href="{{ route('transport.index') }}" class=" @if (Route::currentRouteName() == "transport.index") active @endif ">
-				<span class="icon" data-feather="truck"></span> <span class="ExpandableItem">Transportas</span>
+				<span class="icon" data-feather="truck" id="navIcon"></span> <span class="ExpandableItem">Transportas</span>
 			</a>
 		</div>
 	</div>